Overhaul of Job class, add WorkItem class
Separate out the concept of a "WorkItem" from a "Job".

The reason for this is that we want to collapse lots of similar old-style jobs together into a single place to make managing them easier - notably mex building jobs.

Now we have a single job for which we have a work item for each mass point, instead of a separate job for every mass point.

This makes budgeting of mass spend against all mex construction much easier in the job distribution layer.